tests/wpt/web-platform-tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_min-width-auto.html

<!DOCTYPE html>
<title>flexbox | computed style | min-width: auto</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#min-size-auto">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
html {
	display: flex;
}
body {
	color: red;
	min-width: auto;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("min-width"),
	          "auto");
});
</script>
